+++ sdk/src/main/scala/com/horizen/block/MainchainTransactionInput.scala
@@ -1,6 +1,6 @@
 package com.horizen.block
 
-import com.horizen.utils.{BytesUtils, VarInt}
+import com.horizen.utils.{BytesUtils, CompactSize}
 
 case class MainchainTransactionInput(inputBytes: Array[Byte],
                                      prevTxHash: Array[Byte],
@@ -16,16 +16,16 @@ object MainchainTransactionInput {
     val prevTxHash: Array[Byte] = inputBytes.slice(currentOffset, currentOffset + 32)
     currentOffset += 32
 
-    val prevTxOutputIndex: Int = BytesUtils.getInt(inputBytes, currentOffset)
+    val prevTxOutputIndex: Int = BytesUtils.getReversedInt(inputBytes, currentOffset)
     currentOffset += 4
 
-    val scriptLength: VarInt = BytesUtils.getReversedVarInt(inputBytes, currentOffset)
+    val scriptLength: CompactSize = BytesUtils.getCompactSize(inputBytes, currentOffset)
     currentOffset += scriptLength.size()
 
     val txScript: Array[Byte] = inputBytes.slice(currentOffset, currentOffset + scriptLength.value().intValue())
     currentOffset += scriptLength.value().intValue()
 
-    val sequence: Int = BytesUtils.getInt(inputBytes, currentOffset)
+    val sequence: Int = BytesUtils.getReversedInt(inputBytes, currentOffset)
     currentOffset += 4
 
     MainchainTransactionInput(inputBytes.slice(offset, currentOffset), prevTxHash, prevTxOutputIndex, txScript, sequence)