+++ tools/sctool/src/main/java/com/horizen/CommandProcessor.java
@@ -450,7 +450,7 @@ public class CommandProcessor {
             byte[] scId = Arrays.copyOfRange(infoBytes, offset, offset + 32);
             offset += 32;
 
-            VarInt powDataLength = BytesUtils.getReversedVarInt(infoBytes, offset);
+            CompactSize powDataLength = BytesUtils.getCompactSize(infoBytes, offset);
             offset += powDataLength.size();
 
             String powData = BytesUtils.toHexString(Arrays.copyOfRange(infoBytes, offset, offset + ((int)powDataLength.value() * 8)));
@@ -459,7 +459,7 @@ public class CommandProcessor {
             int mcBlockHeight = BytesUtils.getReversedInt(infoBytes, offset);
             offset += 4;
 
-            VarInt initialCumulativeCommTreeHashLength = BytesUtils.getReversedVarInt(infoBytes, offset);
+            CompactSize initialCumulativeCommTreeHashLength = BytesUtils.getCompactSize(infoBytes, offset);
             offset += initialCumulativeCommTreeHashLength.size();
 
             // Note: we keep this value in Little endian as expected by sc-cryptolib
@@ -476,7 +476,7 @@ public class CommandProcessor {
             SidechainsVersionsManager versionsManager = null;
             if(offset < infoBytes.length) {
                 Map<ByteArrayWrapper, Enumeration.Value> scVersions = new HashMap<>();
-                VarInt scSidechainVersionsLength = BytesUtils.getReversedVarInt(infoBytes, offset);
+                CompactSize scSidechainVersionsLength = BytesUtils.getCompactSize(infoBytes, offset);
                 offset += scSidechainVersionsLength.size();
                 for (int i = 0; i < scSidechainVersionsLength.value(); i++) {
                     byte[] sidechainId = Arrays.copyOfRange(infoBytes, offset, offset + 32);